watcher/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
cachetools==5.1.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
dnspython==2.2.1
frozenlist==1.3.0
google-auth==2.6.6
idna==3.3
kubernetes==23.6.0
multidict==6.0.2
oauthlib==3.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyMySQL==1.0.2
python-dateutil==2.8.2
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
urllib3==1.26.9
websocket-client==1.3.2
yarl==1.7.2
